test(record-manipulation): add unit tests for common transformation primitives
Cover the Constant/Random/Choose suppliers, Strings (hmac/encrypt/decrypt),
Functional, Fail, and Pipeline building blocks used to construct record
masking/generation transformations.

Writing the tests surfaced a few issues in the code under test, fixed here:
- ConstantLongSupplier and RandomLongSupplier took int parameters despite
  being long suppliers, restricting them to int-range values.
- The Random* suppliers accepted invalid min/max combinations silently;
  they now validate bounds in their constructors and throw
  IllegalArgumentException.
- RandomIntSupplier computed an inclusive max via `max + 1`, which could
  overflow at Integer.MAX_VALUE. Switched to an exclusive upper bound and
  renamed its parameters to match the minInclusive/maxExclusive
  convention used by the other Random* suppliers.
- Use.java's demo now logs its output via SLF4J instead of System.out,
  and unused imports were removed from Strings.java and
  JacksonSerializer.java, clearing the module's checkstyle failures.

+++ kroxylicious-filters/kroxylicious-record-manipulation/src/main/java/io/kroxylicious/filter/record/manipulation/common/ConstantLongSupplier.java
@@ -9,9 +9,9 @@ package io.kroxylicious.filter.record.manipulation.common;
 import java.util.function.LongSupplier;
 
 public class ConstantLongSupplier implements LongSupplier {
-    private final int value;
+    private final long value;
 
-    public ConstantLongSupplier(int value) {
+    public ConstantLongSupplier(long value) {
         this.value = value;
     }
